requirements.txt
streamlit
crewai
crewai-tools
langchain-openai
fastapi
uvicorn
google-api-python-client
python-dotenv
requests
langgraph
langchain-mcp-adapters
mcp